+++ src/popup/pages/Send/SignView/index.styled.tsx
@@ -46,7 +46,11 @@ export const StyledContent = styled.div`
   padding: 0px 20px;
 `;
 
-export const StyledWebsiteContainer = styled.div``;
+export const StyledWebsiteContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+`;
 
 export const StyledAccountRow = styled.div`
   margin-top: 20px;